requirement.txt
ttkthemes
python-dotenv
pandas
reportlab
